requirements.txt
Flask==2.2.2
openai_whisper==20230314
soundfile==0.12.1
